feat: Separate User class
